requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.9
torch==2.3.1
torchvision==0.18.1
spandrel==0.4.1
opencv-python-headless==4.12.0.88
pillow==11.1.0
requests==2.32.3
numpy==2.2.3
